requirements.txt
streamlit
python-docx
pdfplumber
pydantic
langchain_core
langchain_community
python-dotenv
langchain_ollama